+++ ChangeLog.rst
@@ -1,12 +1,23 @@
 更新日志
 ========
 
+0.3.14
+------
+
+- [fix] 修复 Author.columns 因为知乎专栏大幅度改版而无法获取到专栏的问题
+- [fix] 修复 Post.column 因为知乎现在允许发布无专栏文章而出错的 bug，对于无专栏的 Post 现在 column 属性返回 None
+- [fix] 修复 Post.slug 因为知乎文章网址变更而无法获取的问题
+- [fix] 修复 Post.column_in_name 因为知乎现在允许发布无专栏文章而出错的 bug，对于无专栏的 Post 现在 column 属性返回 None
+- [fix] 修复因为上述 Bugs 造成的 Author.activities 出错的问题
+- [add] 现在 CollectActType 可以直接从模块顶级 import， 即 from zhihu import CollectActType
+- [fix] 修复 Topic.follower.motto 获取不正确的问题
+
 0.3.13
 ------
 
 - [fix] 修复因知乎将每次点击问题页面「更多」按钮只加载 50 个回答改为 20 个造成的无法获取所有问题的 bug
-- [fix] 获取 post 后，直接调用 post.save 会出错的 bug
-- [fix] 在终端登录时输入的密码改为不可见
+- [fix] 修复获取 post 后，直接调用 post.save 会出错的 bug
+- [change] 在终端登录时输入的密码改为不可见
 - [change] 貌似知乎登录不怎么需要验证码了，现在作为一个可选项，login_in_terminal 和 create_cookies 默认均不要求验证码
 - [fix] 修复因知乎专栏改版，API 地址变更 造成的 Post 类无法使用的问题
 - [add] 增加 client.add_proxy_pool 方法设置代理池